refactor: extract chain management into node/chainman.{h,cpp}
Issue #3030 workstream A4. Moves the chain-management functions out of the
src/main.cpp monolith into a new node/chainman.{h,cpp} module, co-located
(by call graph) with ConnectBlock/DisconnectBlock in validation.cpp:

  ReorganizeChain, DisconnectBlocksBatch, SetBestChain, ForceReorganizeToHash,
  InvalidChainFound, UpdateSyncTime

ReorganizeChain / DisconnectBlocksBatch have no external callers and become
file-static in chainman.cpp; InvalidChainFound / UpdateSyncTime stay internal
(UpdateSyncTime is still called from main.cpp's abandonment path, so it is
declared in chainman.h). SetBestChain and ForceReorganizeToHash are the public
entry points, declared in chainman.h and included by their callers
(validation.cpp, dbwrapper.cpp, rpc/blockchain.cpp).

Supporting linkage changes (move-only, no behavior change):
- The file-static wallet-notify wrapper SetBestChain(const CBlockLocator&) gains
  external linkage (declared in main.h) so the moved SetBestChain can still call
  it; it would otherwise be an unused static in main.cpp.
- g_previous_block_time / g_nTimeBestReceived (external-linkage globals in
  main.cpp, used by both UpdateSyncTime and OutOfSyncByAge) are now declared in
  main.h. g_chain_trust stays defined in main.cpp and is extern'd in chainman.cpp
  (mirroring validation.cpp), per the issue's "keep g_chain_trust as extern".

main.cpp -574, new node/chainman.cpp +~615; registered in CMakeLists
(gridcoin_util). Pure code move; lock annotations and bodies are unchanged.
