+++ src/main.cpp
@@ -16,6 +16,7 @@
 #include "checkpoints.h"
 #include "txdb.h"
 #include "init.h"
+#include "node/chainman.h"
 #include "node/ui_interface.h"
 #include "gridcoin/beacon.h"
 #include "gridcoin/claim.h"
@@ -181,8 +182,9 @@ void static EraseFromWallets(uint256 hash)
         pwallet->EraseFromWallet(hash);
 }
 
-// notify wallets about a new best chain
-void static SetBestChain(const CBlockLocator& loc)
+// notify wallets about a new best chain. External linkage: also called from
+// node/chainman.cpp's SetBestChain after the chain-management move (issue #3030 A4).
+void SetBestChain(const CBlockLocator& loc)
     EXCLUSIVE_LOCKS_REQUIRED(cs_setpwalletRegistered)
 {
     for (auto const& pwallet : setpwalletRegistered)
@@ -635,39 +637,6 @@ bool IsInitialBlockDownload()
             pindexBest->GetBlockTime() <  GetAdjustedTime() - 8 * 60 * 60);
 }
 
-void static InvalidChainFound(CBlockIndex* pindexNew) EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-{
-    arith_uint256 nBestInvalidBlockTrust = pindexNew->GetBlockTrust();
-    arith_uint256 nBestBlockTrust = pindexBest->GetBlockTrust();
-
-    LogPrintf("InvalidChainFound: invalid block=%s  height=%d  trust=%s  blocktrust=%" PRId64 "  date=%s",
-      pindexNew->GetBlockHash().ToString().substr(0,20),
-      pindexNew->nHeight,
-      g_chain_trust.GetTrust(pindexNew).ToString(),
-      nBestInvalidBlockTrust.GetLow64(),
-      DateTimeStrFormat("%x %H:%M:%S", pindexNew->GetBlockTime()));
-    LogPrintf("InvalidChainFound:  current best=%s  height=%d  trust=%s  blocktrust=%" PRId64 "  date=%s",
-      hashBestChain.ToString().substr(0,20),
-      nBestHeight,
-      g_chain_trust.Best().ToString(),
-      nBestBlockTrust.GetLow64(),
-      DateTimeStrFormat("%x %H:%M:%S", pindexBest->GetBlockTime()));
-}
-
-
-static void UpdateSyncTime(const CBlockIndex* const pindexBest)
-{
-    if (pindexBest && pindexBest->pprev) {
-        g_previous_block_time.store(pindexBest->pprev->GetBlockTime());
-    } else {
-        g_previous_block_time.store(0);
-    }
-
-    if (!OutOfSyncByAge()) {
-        g_nTimeBestReceived.store(GetAdjustedTime());
-    }
-}
-
 bool OutOfSyncByAge()
 {
     // Assume we are out of sync if the current block age is 10
@@ -722,549 +691,6 @@ GRC::SuperblockPtr CBlock::GetSuperblock(const CBlockIndex* const pindex) const
     superblock.Rebind(pindex);
 
     return superblock;
-}
-
-bool ReorganizeChain(CTxDB& txdb, unsigned &cnt_dis, unsigned &cnt_con, CBlock &blockNew, CBlockIndex* pindexNew);
-bool ForceReorganizeToHash(uint256 NewHash)
-{
-    LOCK(cs_main);
-    CTxDB txdb;
-
-    auto mapItem = mapBlockIndex.find(NewHash);
-    if (mapItem == mapBlockIndex.end()) {
-        return error("%s: failed to find requested block in block index", __func__);
-    }
-
-    CBlockIndex* pindexCur = pindexBest;
-    CBlockIndex* pindexNew = mapItem->second;
-    LogPrintf("INFO: %s: current best height %i hash %s "
-              " Target height %i hash %s",
-              __func__,
-              pindexCur->nHeight,pindexCur->GetBlockHash().GetHex(),
-              pindexNew->nHeight,pindexNew->GetBlockHash().GetHex());
-
-    CBlock blockNew;
-    if (!ReadBlockFromDisk(blockNew, pindexNew, Params().GetConsensus())) {
-        return error("%s: Fatal Error while reading new best block.", __func__);
-    }
-
-    const arith_uint256 previous_chain_trust = g_chain_trust.Best();
-    unsigned cnt_dis = 0;
-    unsigned cnt_con = 0;
-    bool success = false;
-
-    // g_reorg_in_progress is set in ReorganizeChain, because ReorganizeChain determines whether this is a trivial
-    // reorg, where we are just adding a new block to the head of the current chain, in which case we do not
-    // want the flag to be set.
-    success = ReorganizeChain(txdb, cnt_dis, cnt_con, blockNew, pindexNew);
-
-    if (g_chain_trust.Best() < previous_chain_trust) {
-        LogPrintf("WARN: %s: Chain trust is now less than before!", __func__);
-    }
-
-    // g_reorg_in_progress is set in ReorganizeChain, but cleared by the caller. It is debatable whether this should
-    // be cleared here if g_chain_trust.Best() < previous_chain_trust.
-    g_reorg_in_progress = false;
-
-    if (!success) {
-        return error("%s: Fatal Error while setting best chain.", __func__);
-    }
-
-    AskForOutstandingBlocks(uint256());
-    LogPrintf("INFO %s: success! height %d hash %s",
-              __func__,
-              pindexBest->nHeight,
-              pindexBest->GetBlockHash().GetHex());
-
-    return true;
-}
-
-bool DisconnectBlocksBatch(CTxDB& txdb, list<CTransaction>& vResurrect, unsigned& cnt_dis, CBlockIndex* pcommon) EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-{
-    set<string> vRereadCPIDs;
-
-    GRC::RegistryBookmarks registries;
-
-    // Count superblocks crossed during this disconnect. BeaconRegistry::Deactivate
-    // is called per-SB below and is fidelity-correct for the SINGLE most recent SB
-    // (it uses m_expired_pending which only carries the latest SB's expired set --
-    // see beacon.cpp:1265-1273). When the disconnect crosses 2+ SBs, expired-pending
-    // beacons from the prior SBs are lost; the recovery is an in-line beacon
-    // registry rebuild after the disconnect loop completes (call site below at the
-    // sb_cross_count >= 2 check). The full rationale for in-line vs deferred
-    // rebuild and why single-SB reorgs are NOT rebuilt lives at that call site.
-    int sb_cross_count = 0;
-
-    while(pindexBest != pcommon)
-    {
-        if(!pindexBest->pprev)
-            return error("DisconnectBlocksBatch: attempt to reorganize beyond genesis"); /*fatal*/
-
-        LogPrint(BCLog::LogFlags::VERBOSE, "DisconnectBlocksBatch: %s",pindexBest->GetBlockHash().GetHex());
-
-        CBlock block;
-        if (!ReadBlockFromDisk(block, pindexBest, Params().GetConsensus()))
-            return error("DisconnectBlocksBatch: ReadFromDisk for disconnect failed"); /*fatal*/
-        if (!DisconnectBlock(block, txdb, pindexBest))
-            return error("DisconnectBlocksBatch: DisconnectBlock %s failed", pindexBest->GetBlockHash().ToString().c_str()); /*fatal*/
-
-        // disconnect from memory
-        assert(!pindexBest->pnext);
-        if (pindexBest->pprev)
-            pindexBest->pprev->pnext = nullptr;
-
-        // Queue memory transactions to resurrect.
-        // We only do this for blocks after the last checkpoint (reorganisation before that
-        // point should only happen with -reindex/-loadblock, or a misbehaving peer).
-        for (auto const& tx : boost::adaptors::reverse(block.vtx))
-            if (!(tx.IsCoinBase() || tx.IsCoinStake()) && pindexBest->nHeight > Params().Checkpoints().GetHeight())
-                vResurrect.push_front(tx);
-
-        // TODO: Implement flag in CBlockIndex for mrcs?
-        if (pindexBest->IsUserCPID() || !pindexBest->m_mrc_researchers.empty()) {
-            // The user has no longer staked this block.
-            GRC::Tally::ForgetRewardBlock(pindexBest);
-        }
-
-        if (pindexBest->IsSuperblock()) {
-            // Revert beacon activations which were done by the superblock to revert and resurrect pending records
-            // for the reverted activations. This is safe to do before the transactional level reverts with beacon
-            // contracts, because any beacon that is activated CANNOT have been a new advertisement in the superblock
-            // itself. It would not be verified. AND if the beacon is a renewal, it would never be in the activation list
-            // for a superblock. We call GetBeaconRegistry directly here, because the IHandler class does not have
-            // a virtual method that corresponds to this call, as it is only relevant to beacons.
-            GRC::GetBeaconRegistry().Deactivate(pindexBest->GetBlockHash());
-
-            // Count it so we can decide after the loop whether to flag a beacon-registry rebuild
-            // for the next startup (the Deactivate-via-m_expired_pending path is only correct
-            // for one SB at a time; see the comment above the loop).
-            ++sb_cross_count;
-
-            GRC::Quorum::PopSuperblock(pindexBest);
-            GRC::Quorum::LoadSuperblockIndex(pindexBest->pprev);
-
-            if (pindexBest->nVersion >= 11 && !GRC::Tally::RevertSuperblock()) {
-                return false;
-            }
-        }
-
-        if (pindexBest->nHeight > nGrandfather && pindexBest->nVersion <= 10) {
-            GRC::Quorum::ForgetVote(pindexBest);
-        }
-
-        // Delete beacons, scraper entries, protocol entries, projects, polls and votes from contracts
-        // in disconnected blocks.
-        if (pindexBest->IsContract())
-        {
-            // Skip coinbase and coinstake transactions:
-            for (auto tx = std::next(block.vtx.begin(), 2), end = block.vtx.end();
-                tx != end;
-                ++tx)
-            {
-                // This reverts contracts for those contract types which have handlers that properly handle
-                // contract level reversions.
-                for (const auto& contract : tx->GetContracts())
-                {
-                    if (GRC::RegistryBookmarks::IsRegistryRevertCapable(contract.m_type.Value())) {
-                        const GRC::ContractContext contract_context(contract, *tx, pindexBest);
-
-                        GRC::RegistryBookmarks::GetRegistryWithRevert(contract.m_type.Value()).Revert(contract_context);
-                    }
-                }
-            }
-        }
-
-        // New best block
-        cnt_dis++;
-        pindexBest = pindexBest->pprev;
-        hashBestChain = pindexBest->GetBlockHash();
-        nBestHeight = pindexBest->nHeight;
-        g_chain_trust.SetBest(pindexBest);
-
-        UpdateSyncTime(pindexBest);
-
-        if (!txdb.WriteHashBestChain(pindexBest->GetBlockHash()))
-            return error("DisconnectBlocksBatch: WriteHashBestChain failed"); /*fatal*/
-
-    }
-
-    /* fix up after disconnecting, prepare for new blocks */
-    if (cnt_dis > 0)
-    {
-        // Resurrect memory transactions that were in the disconnected branch
-        for( CTransaction& tx : vResurrect) {
-            CValidationState resurrect_state;
-            AcceptToMemoryPool(mempool, tx, resurrect_state, nullptr);
-        }
-
-        if (!txdb.TxnCommit())
-            return error("DisconnectBlocksBatch: TxnCommit failed"); /*fatal*/
-
-        // Record new best height (the common block) in the registries that have a backing DB. This is important
-        // to ensure that if the wallet is shutdown, on the next start, the contract replay (if any) is done from
-        // the correct height.
-        registries.UpdateRegistryBlockHeights(pindexBest->nHeight);
-
-        // Replaying contracts after a block disconnection is no longer needed, as all contract types that have handlers
-        // that operate at the tx/contract level have fully implemented reversion.
-        //GRC::ReplayContracts(pindexBest);
-
-        // Tally research averages.
-        if(IsV9Enabled_Tally(nBestHeight) && !IsV11Enabled(nBestHeight)) {
-            assert(GRC::Tally::IsLegacyTrigger(nBestHeight));
-            GRC::Tally::LegacyRecount(pindexBest);
-        }
-
-        // If the reorg crossed two or more superblock boundaries, the per-SB Deactivate calls
-        // above could not fully resurrect expired-pending beacons from the prior SBs
-        // (m_expired_pending only carries the most recent SB's set -- the limitation
-        // acknowledged at beacon.cpp:1265-1273). Rebuild the beacon registry in-line
-        // BEFORE returning to the reconnect side of ReorganizeChain: the rebuild walks
-        // from V11_height to the current (common-ancestor) tip, leaving the registry
-        // correctly reflecting that ancestor's state, and the subsequent per-block
-        // Activate() calls on the reconnect side then bring the registry up to the
-        // new tip.
-        //
-        // Doing this in-line (rather than deferring to a flag picked up on next
-        // startup) trades a few seconds of frozen-wallet time for closing the fork
-        // window: with the deferred approach, any block validated between this reorg
-        // and the next restart could consult the broken registry and diverge from
-        // healthy peers. In-line rebuild is bounded by the chain walk from V11 to
-        // tip, which is dominated by block-index iteration -- on SSD typically
-        // single-digit seconds. See doc/block_corruption_recovery_design.md.
-        //
-        // Single-SB reorgs are NOT rebuilt: the existing Deactivate path is
-        // fidelity-correct for that case and the rebuild would be wasted work on
-        // the most common case.
-        if (sb_cross_count >= 2) {
-            LogPrintf("WARN: %s: reorg disconnected %d superblock(s); beacon registry expired-pending "
-                      "fidelity is degraded for the prior SB(s). Rebuilding beacon registry in-line "
-                      "to maintain consensus.",
-                      __func__, sb_cross_count);
-            GRC::RebuildBeaconRegistry();
-        }
-    }
-
-    return true;
-}
-
-bool ReorganizeChain(CTxDB& txdb, unsigned &cnt_dis, unsigned &cnt_con, CBlock &blockNew, CBlockIndex* pindexNew)
-EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-{
-    assert(pindexNew);
-    assert(pindexNew->GetBlockHash()==blockNew.GetHash(true));
-    /* note: it was already determined that this chain is better than current best */
-    /* assert(pindexNew->nChainTrust > nBestChainTrust); but may be overridden by command */
-    assert(!pindexGenesisBlock == !pindexBest);
-
-    list<CTransaction> vResurrect;
-    list<CBlockIndex*> vConnect;
-    set<string> vRereadCPIDs;
-
-    /* find fork point */
-    CBlockIndex* pcommon = nullptr;
-
-    if (pindexGenesisBlock) {
-        pcommon = pindexNew;
-        // The trivial reorg that most often happens here is where pcommon IS pindexBest, which
-        // means effectively we are simply adding a new block to the end of the current chain.
-        // Notice the logic of the while statement allows one traversal in that case... When
-        // pcommon = pindexNew then pcommon->pnext will be nullptr. It will also satisfy
-        // pcommon != pindexBest. Then the pointer moves back one block, which in the trivial
-        // case will be pindexBest and the while loop ends with pcommon = pindexBest.
-        while (pcommon->pnext == nullptr && pcommon != pindexBest) {
-            pcommon = pcommon->pprev;
-
-            if (!pcommon) {
-                return error("%s: unable to find fork root", __func__);
-            }
-        }
-
-        // Blocks version 11+ do not use the legacy tally system triggered by
-        // block height intervals:
-        //
-        if (pcommon->nVersion <= 10 && pcommon != pindexBest) {
-            pcommon = GRC::Tally::FindLegacyTrigger(pcommon);
-
-            if (!pcommon) {
-                return error("%s: unable to find fork root with tally point", __func__);
-            }
-        }
-
-        if (pcommon != pindexBest || pindexNew->pprev != pcommon) {
-            // This is set for the benefit of other threads, such as the GUI or rpc for the poll registry,
-            // that do work while not locking cs_main.
-            g_reorg_in_progress = true;
-
-            LogPrintf("INFO: %s: from {%s %d}, common {%s %d}, to {%s %d}, disconnecting %d, connecting %d blocks",
-                      __func__,
-                      pindexBest->GetBlockHash().GetHex().c_str(), pindexBest->nHeight,
-                      pcommon->GetBlockHash().GetHex().c_str(), pcommon->nHeight,
-                      pindexNew->GetBlockHash().GetHex().c_str(), pindexNew->nHeight,
-                      pindexBest->nHeight - pcommon->nHeight,
-                      pindexNew->nHeight - pcommon->nHeight);
-        }
-    }
-
-    /* disconnect blocks (in non-trivial condition) */
-    if (pcommon != pindexBest) {
-        if (!txdb.TxnBegin()) {
-            return error("%s: TxnBegin failed", __func__);
-        }
-        if (!DisconnectBlocksBatch(txdb, vResurrect, cnt_dis, pcommon)) {
-            error("%s: DisconnectBlocksBatch() failed. This is a fatal error. The chain index may be corrupt. Aborting. "
-                  "Please reindex the chain and restart.", __func__);
-            exit(1); //todo
-        }
-
-        int nMismatchSpent;
-        int64_t nBalanceInQuestion;
-        pwalletMain->FixSpentCoins(nMismatchSpent, nBalanceInQuestion);
-    }
-
-    if (LogInstance().WillLogCategory(BCLog::LogFlags::VERBOSE) && cnt_dis > 0) {
-
-        LogPrintf("INFO: %s: disconnected %d blocks", __func__, cnt_dis);
-    }
-
-    for (CBlockIndex *p = pindexNew; p != pcommon; p=p->pprev) {
-        vConnect.push_front(p);
-    }
-
-    /* Connect blocks */
-    for (auto const pindex : vConnect) {
-        CBlock block_load;
-        CBlock &block = (pindex==pindexNew)? blockNew : block_load;
-
-        if (pindex != pindexNew) {
-            if (!ReadBlockFromDisk(block, pindex, Params().GetConsensus())) {
-                return error("%s: ReadFromDisk for connect failed", __func__);
-            }
-
-            assert(pindex->GetBlockHash()==block.GetHash(true));
-        } else {
-            assert(pindex == pindexNew);
-            assert(pindexNew->GetBlockHash() == block.GetHash(true));
-            assert(pindexNew->GetBlockHash() == blockNew.GetHash(true));
-        }
-
-        uint256 hash = block.GetHash(true);
-
-        LogPrint(BCLog::LogFlags::VERBOSE, "INFO: %s: connect %s", __func__, hash.ToString());
-
-        if (!txdb.TxnBegin()) {
-            return error("%s: TxnBegin failed", __func__);
-        }
-
-        {
-            // This lock protects the time period between the GridcoinConnectBlock, which also connects validated transaction
-            // contracts and causes contract handlers to fire, and the committing of the txindex changes to disk. Any contract
-            // handlers that generate signals whose downstream handlers make use of transaction data on disk via leveldb (txdb)
-            // on another thread need to take this lock to ensure that the write to leveldb and the access of the transaction data
-            // by the signal handlers is appropriately serialized.
-            LOCK(cs_tx_val_commit_to_disk);
-            LogPrint(BCLog::LogFlags::VOTE, "INFO: %s: cs_tx_val_commit_to_disk locked", __func__);
-
-            if (pindexGenesisBlock == nullptr) {
-                const uint256 expected_genesis = Params().IsMockableChain()
-                    ? (hashGenesisBlockRegTest.IsNull() ? hash : hashGenesisBlockRegTest)
-                    : (!fTestNet ? hashGenesisBlock : hashGenesisBlockTestNet);
-                if (hash != expected_genesis) {
-                    txdb.TxnAbort();
-                    return error("%s: genesis block hash does not match", __func__);
-                }
-
-                pindexGenesisBlock = pindex;
-            } else {
-                assert(pindex->GetBlockHash()==block.GetHash(true));
-                assert(pindex->pprev == pindexBest);
-
-                CValidationState connect_state;
-                if (!ConnectBlock(block, connect_state, txdb, pindex, false)) {
-                    txdb.TxnAbort();
-                    error("%s: ConnectBlock %s failed, Previous block %s",
-                          __func__,
-                          hash.ToString().c_str(),
-                          pindex->pprev->GetBlockHash().ToString());
-                    InvalidChainFound(pindex);
-                    return false;
-                }
-            }
-
-            // Delete redundant memory transactions
-            for (auto const& tx : block.vtx) {
-                mempool.remove(tx);
-                mempool.removeConflicts(tx);
-            }
-
-            // Remove stale MRCs in the mempool that are not in this new block. Remember the MRCs were initially validated in
-            // AcceptToMemoryPool. Here we just need to do a staleness check.
-            std::vector<CTransaction> to_be_erased;
-
-            for (const auto& [_, pool_tx] : mempool.mapTx) {
-                for (const auto& pool_tx_contract : pool_tx.GetContracts()) {
-                    if (pool_tx_contract.m_type == GRC::ContractType::MRC) {
-                        GRC::MRC pool_tx_mrc = pool_tx_contract.CopyPayloadAs<GRC::MRC>();
-
-                        if (pool_tx_mrc.m_last_block_hash != hashBestChain) {
-                            to_be_erased.push_back(pool_tx);
-                        }
-                    }
-                }
-            }
-
-            // TODO: Additional mempool removals for generic transactions based on txns...
-            // that satisfy lock time requirements,
-            // that are at least 30m old,
-            // that have been broadcast at least once min 5m ago,
-            // that had at least 45s to go in to the last block,
-            // and are still not in the txdb? (for the wallet itself, not mempool.)
-
-            for (const auto& tx : to_be_erased) {
-                LogPrintf("%s: Erasing stale transaction %s from mempool and wallet.", __func__, tx.GetHash().ToString());
-                mempool.remove(tx);
-                // If this transaction was in this wallet (i.e. erasure successful), then send signal for GUI.
-                if (pwalletMain->EraseFromWallet(tx.GetHash())) {
-                    pwalletMain->NotifyTransactionChanged(pwalletMain, tx.GetHash(), CT_DELETED);
-                }
-            }
-
-            // Clean up spent outputs in wallet that are now not spent if mempool transactions erased above. This
-            // is ugly and heavyweight and should be replaced when the upstream wallet code is ported. Unlike the
-            // repairwallet rpc, this is silent.
-            if (!to_be_erased.empty()) {
-                int nMisMatchFound = 0;
-                CAmount nBalanceInQuestion = 0;
-
-                pwalletMain->FixSpentCoins(nMisMatchFound, nBalanceInQuestion);
-            }
-
-            if (!txdb.WriteHashBestChain(pindex->GetBlockHash())) {
-                txdb.TxnAbort();
-                return error("%s: WriteHashBestChain failed", __func__);
-            }
-
-            // Make sure it's successfully written to disk before changing memory structure
-            if (!txdb.TxnCommit()) {
-                return error("%s: TxnCommit failed", __func__);
-            }
-
-            LogPrint(BCLog::LogFlags::VOTE, "INFO: %s: cs_tx_val_commit_to_disk unlocked", __func__);
-        }
-
-        // Add to current best branch
-        if (pindex->pprev) {
-            assert(!pindex->pprev->pnext);
-            pindex->pprev->pnext = pindex;
-        }
-
-        // update best block
-        hashBestChain = hash;
-        pindexBest = pindex;
-        nBestHeight = pindexBest->nHeight;
-        g_chain_trust.SetBest(pindexBest);
-        cnt_con++;
-
-        UpdateSyncTime(pindexBest);
-
-        if (IsV9Enabled_Tally(nBestHeight)
-            && !IsV11Enabled(nBestHeight)
-            && GRC::Tally::IsLegacyTrigger(nBestHeight)) {
-
-            GRC::Tally::LegacyRecount(pindexBest);
-        }
-    }
-
-    if (LogInstance().WillLogCategory(BCLog::LogFlags::VERBOSE) && (cnt_dis > 0 || cnt_con > 1)) {
-        LogPrintf("INFO %s: Disconnected %d and connected %d blocks.",__func__, cnt_dis, cnt_con);
-    }
-
-    return true;
-}
-
-bool SetBestChain(CTxDB& txdb, CBlock &blockNew, CBlockIndex* pindexNew) EXCLUSIVE_LOCKS_REQUIRED(cs_main)
-{
-    unsigned cnt_dis = 0;
-    unsigned cnt_con = 0;
-    bool success = false;
-    const auto origBestIndex = pindexBest;
-    const arith_uint256 previous_chain_trust = g_chain_trust.Best();
-
-    // g_reorg_in_progress is set in ReorganizeChain, because ReorganizeChain determines whether this is a trivial
-    // reorg, where we are just adding a new block to the head of the current chain, in which case we do not
-    // want the flag to be set.
-    success = ReorganizeChain(txdb, cnt_dis, cnt_con, blockNew, pindexNew);
-
-    if (previous_chain_trust > g_chain_trust.Best()) {
-        LogPrintf("INFO: %s: Reorganize caused lower chain trust than before. Reorganizing back.", __func__);
-
-        CBlock origBlock;
-
-        if (!ReadBlockFromDisk(origBlock, origBestIndex, Params().GetConsensus())) {
-            return error("%s: Fatal Error while reading original best block", __func__);
-        }
-
-        success = ReorganizeChain(txdb, cnt_dis, cnt_con, origBlock, origBestIndex);
-    }
-
-    if (!success) {
-        return false;
-    }
-
-    // g_reorg_in_progress is set in ReorganizeChain, but cleared by the caller, because as above it can
-    // be called more than once as part of what is really a single reorg.
-    g_reorg_in_progress = false;
-
-    /* Fix up after block connecting */
-
-    // Update best block in wallet (so we can detect restored wallets)
-    bool fIsInitialDownload = IsInitialBlockDownload();
-
-    if (!fIsInitialDownload) {
-        const CBlockLocator locator(pindexNew);
-        // Canonical order: cs_main (held by SetBestChain) -> cs_setpwalletRegistered -> cs_wallet.
-        LOCK(cs_setpwalletRegistered);
-        ::SetBestChain(locator);
-    }
-
-    if (LogInstance().WillLogCategory(BCLog::LogFlags::VERBOSE)) {
-        LogPrintf("INFO: %s: new best block {%s %d}, trust=%s, date=%s",
-                  __func__,
-                  hashBestChain.ToString(), nBestHeight,
-                  g_chain_trust.Best().ToString(),
-                  DateTimeStrFormat("%x %H:%M:%S", pindexBest->GetBlockTime()));
-    } else {
-        LogPrintf("INFO: %s: new best block {%s %d}",
-                  __func__,
-                  hashBestChain.ToString(),
-                  nBestHeight);
-    }
-
-    #if HAVE_SYSTEM
-    std::string strCmd = gArgs.GetArg("-blocknotify", "");
-    if (!fIsInitialDownload && !strCmd.empty())
-    {
-        ReplaceAll(strCmd, "%s", hashBestChain.GetHex());
-        boost::thread t(runCommand, strCmd); // thread runs free
-    }
-    #endif
-
-    // Notify the validation-signal layer that the chain tip advanced. The UI
-    // bridge registered in init.cpp re-emits uiInterface.NotifyBlocksChanged()
-    // for the Qt models, so GUI block notifications now flow through the
-    // validation interface (issue #3030, workstream B3).
-    //
-    // pindexFork is passed as nullptr: the true reorg fork point is the common
-    // ancestor (pcommon, computed in ReorganizeChain) and is not plumbed up to
-    // this layer, so origBestIndex (the previous tip) would be wrong for any
-    // non-trivial reorg. No current subscriber reads pindexFork (the UI bridge
-    // ignores it), so nullptr is an honest "not supplied" rather than a
-    // misleading value; a future fork-point-dependent subscriber (e.g. the
-    // deferred PeerManager) should thread pcommon through instead. Tracked in
-    // issue #3104.
-    GetMainSignals().UpdatedBlockTip(pindexNew, nullptr, fIsInitialDownload);
-
-    return GridcoinServices();
 }
 
 arith_uint256 CBlockIndex::GetBlockTrust() const
